requirements.txt
Bottleneck @ file:///C:/ci/bottleneck_1648010904582/work
efficientnet-pytorch==0.7.1
joblib @ file:///tmp/build/80754af9/joblib_1635411271373/work
mkl-fft==1.3.1
mkl-random @ file:///C:/ci/mkl_random_1626186184308/work
mkl-service==2.4.0
numexpr @ file:///C:/ci/numexpr_1656071335867/work
numpy @ file:///C:/ci/numpy_and_numpy_base_1652784039997/work
packaging @ file:///tmp/build/80754af9/packaging_1637314298585/work
pandas @ file:///C:/ci/pandas_1650373458095/work
Pillow==9.2.0
pyparsing @ file:///tmp/build/80754af9/pyparsing_1635766073266/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytz @ file:///C:/Windows/TEMP/abs_90eacd4e-8eff-491e-b26e-f707eba2cbe1ujvbhqz1/croots/recipe/pytz_1654762631027/work
scikit-learn @ file:///C:/ci/scikit-learn_1658344472767/work
scipy @ file:///C:/ci/scipy_1641555170412/work
six @ file:///tmp/build/80754af9/six_1644875935023/work
sklearn==0.0
threadpoolctl @ file:///Users/ktietz/demo/mc3/conda-bld/threadpoolctl_1629802263681/work
torch==1.9.0
torchaudio==0.9.0
torchsummary==1.5.1
torchvision==0.10.0
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1656706066251/work
